KNN Explainer 5/6 tests passed Divided Threshold into another class Renamed threshold_test.py
